+++ packages/engine/Source/Scene/GlobeDepth.js
@@ -2,6 +2,7 @@ import BoundingRectangle from "../Core/BoundingRectangle.js";
 import Color from "../Core/Color.js";
 import defined from "../Core/defined.js";
 import destroyObject from "../Core/destroyObject.js";
+import PixelFormat from "../Core/PixelFormat.js";
 import ClearCommand from "../Renderer/ClearCommand.js";
 import FramebufferManager from "../Renderer/FramebufferManager.js";
 import PixelDatatype from "../Renderer/PixelDatatype.js";
@@ -21,12 +22,23 @@ import StencilOperation from "./StencilOperation.js";
  */
 function GlobeDepth() {
   this._picking = false;
+  this._snapping = false;
   this._numSamples = 1;
   this._tempCopyDepthTexture = undefined;
 
   this._pickColorFramebuffer = new FramebufferManager({
     depthStencil: true,
     supportsDepthTexture: true,
+  });
+  // Float twin of the pick color framebuffer, used during a snapping pass
+  // (see Scene#snap). RGBA32F so each pixel can carry the snap payload
+  // (uint32 pick ID, isEdge flag, eye-space depth). GPU resources are only
+  // allocated on first use (FramebufferManager allocates lazily in update).
+  this._snapColorFramebuffer = new FramebufferManager({
+    depthStencil: true,
+    supportsDepthTexture: true,
+    pixelDatatype: PixelDatatype.FLOAT,
+    pixelFormat: PixelFormat.RGBA,
   });
   this._outputFramebuffer = new FramebufferManager({
     depthStencil: true,
@@ -61,6 +73,9 @@ function GlobeDepth() {
 Object.defineProperties(GlobeDepth.prototype, {
   colorFramebufferManager: {
     get: function () {
+      if (this._snapping) {
+        return this._snapColorFramebuffer;
+      }
       return this._picking
         ? this._pickColorFramebuffer
         : this._outputFramebuffer;
@@ -82,6 +97,14 @@ Object.defineProperties(GlobeDepth.prototype, {
     },
     set: function (value) {
       this._picking = value;
+    },
+  },
+  snapping: {
+    get: function () {
+      return this._snapping;
+    },
+    set: function (value) {
+      this._snapping = value;
     },
   },
 });
@@ -260,7 +283,9 @@ GlobeDepth.prototype.update = function (
       : PixelDatatype.FLOAT
     : PixelDatatype.UNSIGNED_BYTE;
   this._numSamples = numSamples;
-  if (this.picking) {
+  if (this.snapping) {
+    this._snapColorFramebuffer.update(context, width, height);
+  } else if (this.picking) {
     this._pickColorFramebuffer.update(context, width, height);
   } else {
     this._outputFramebuffer.update(
@@ -377,6 +402,7 @@ GlobeDepth.prototype.isDestroyed = function () {
 
 GlobeDepth.prototype.destroy = function () {
   this._pickColorFramebuffer.destroy();
+  this._snapColorFramebuffer.destroy();
   this._outputFramebuffer.destroy();
   this._copyDepthFramebuffer.destroy();
   this._tempCopyDepthFramebuffer.destroy();